Split MCP tests by domain
